Move records to dedicated packages
