+++ asyncapi-core/src/main/java/io/ballerina/asyncapi/core/model/channel/HttpChannelBindings.java
@@ -1,0 +1,5 @@
+package io.ballerina.asyncapi.core.model.channel;
+
+/** HTTP Channel Binding Object. Reserved for future use; has no defined properties. */
+public record HttpChannelBindings() {
+}
