+++ asyncapi-core/src/main/java/io/ballerina/asyncapi/core/model/message/WsMessageBindings.java
@@ -1,0 +1,5 @@
+package io.ballerina.asyncapi.core.model.message;
+
+/** WebSocket Message Binding Object. Reserved for future use; has no defined properties. */
+public record WsMessageBindings() {
+}
