+++ asyncapi-core/src/main/java/io/ballerina/asyncapi/core/model/operation/WsOperationBindings.java
@@ -1,0 +1,5 @@
+package io.ballerina.asyncapi.core.model.operation;
+
+/** WebSocket Operation Binding Object. Reserved for future use; has no defined properties. */
+public record WsOperationBindings() {
+}
