+++ asyncapi-core/src/main/java/io/ballerina/asyncapi/core/model/server/HttpServerBindings.java
@@ -1,0 +1,5 @@
+package io.ballerina.asyncapi.core.model.server;
+
+/** HTTP Server Binding Object. Reserved for future use; has no defined properties. */
+public record HttpServerBindings() {
+}
