+++ asyncapi-core/src/main/java/io/ballerina/asyncapi/core/model/server/WsServerBindings.java
@@ -1,0 +1,5 @@
+package io.ballerina.asyncapi.core.model.server;
+
+/** WebSocket Server Binding Object. Reserved for future use; has no defined properties. */
+public record WsServerBindings() {
+}
