+++ asyncapi-generator/build.gradle
@@ -2,19 +2,10 @@ plugins {
     id 'java'
 }
 
-group = 'io.ballerina'
-version = '0.12.1-SNAPSHOT'
-
 repositories {
     mavenCentral()
 }
 
 dependencies {
-    testImplementation platform('org.junit:junit-bom:5.10.0')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-}
 
-test {
-    useJUnitPlatform()
 }
